decoupled client and dataset creation; moved dataset preparation aux functions to `dataset_preparation.py`; other minor changes
