+++ baselines/flwr_baselines/publications/fedprox_mnist/dataset.py
@@ -4,12 +4,10 @@
 from typing import List, Optional, Tuple
 
 from omegaconf import DictConfig
-import numpy as np
 import torch
-import torchvision.transforms as transforms
-from torch.utils.data import ConcatDataset, DataLoader, Dataset, Subset, random_split
-from torchvision.datasets import MNIST
+from torch.utils.data import DataLoader, random_split
 
+from flwr_baselines.publications.fedprox_mnist.dataset_preparation import _partition_data
 
 def load_datasets(  # pylint: disable=too-many-arguments
     num_clients: int,
@@ -17,7 +15,7 @@ def load_datasets(  # pylint: disable=too-many-arguments
     val_ratio: float = 0.1,
     batch_size: Optional[int] = 32,
     seed: Optional[int] = 42,
-) -> Tuple[DataLoader, DataLoader, DataLoader]:
+) -> Tuple[List[DataLoader], List[DataLoader], DataLoader]:
     """Creates the dataloaders to be fed into the model.
 
     Parameters
@@ -53,233 +51,3 @@ def load_datasets(  # pylint: disable=too-many-arguments
         trainloaders.append(DataLoader(ds_train, batch_size=batch_size, shuffle=True, num_workers=2))
         valloaders.append(DataLoader(ds_val, batch_size=batch_size))
     return trainloaders, valloaders, DataLoader(testset, batch_size=batch_size)
-
-
-def _download_data() -> Tuple[Dataset, Dataset]:
-    """Downloads (if necessary) and returns the MNIST dataset.
-
-    Returns
-    -------
-    Tuple[MNIST, MNIST]
-        The dataset for training and the dataset for testing MNIST.
-    """
-    transform = transforms.Compose(
-        [transforms.ToTensor(), transforms.Normalize((0.1307,), (0.3081,))]
-    )
-    trainset = MNIST("./dataset", train=True, download=True, transform=transform)
-    testset = MNIST("./dataset", train=False, download=True, transform=transform)
-    return trainset, testset
-
-
-def _partition_data(
-    num_clients,
-    iid: Optional[bool] = False,
-    power_law: Optional[bool] = True,
-    balance: Optional[bool] = False,
-    seed: Optional[int] = 42,
-) -> Tuple[List[Dataset], Dataset]:
-    """Split training set into iid or non iid partitions to simulate the
-    federated setting.
-
-    Parameters
-    ----------
-    num_clients : int
-        The number of clients that hold a part of the data
-    iid : bool, optional
-        Whether the data should be independent and identically distributed between
-        the clients or if the data should first be sorted by labels and distributed by chunks
-        to each client (used to test the convergence in a worst case scenario), by default False
-    power_law: bool, optional
-        Whether to follow a power-law distribution when assigning number of samples
-        for each client, defaults to True
-    balance : bool, optional
-        Whether the dataset should contain an equal number of samples in each class,
-        by default False
-    seed : int, optional
-        Used to set a fix seed to replicate experiments, by default 42
-
-    Returns
-    -------
-    Tuple[List[Dataset], Dataset]
-        A list of dataset for each client and a single dataset to be use for testing the model.
-    """
-    trainset, testset = _download_data()
-    partition_size = int(len(trainset) / num_clients)
-    lengths = [partition_size] * num_clients
-    if iid:
-        datasets = random_split(trainset, lengths, torch.Generator().manual_seed(seed))
-    else:
-        if power_law:
-            trainset_sorted = _sort_by_class(trainset)
-            datasets = _power_law_split(trainset_sorted, num_partitions=num_clients,
-                                        num_labels_per_partition=2, min_data_per_partition=10,
-                                        mean=0.0, sigma=2.0)
-        else:
-            if balance:
-                trainset = _balance_classes(trainset, seed)
-                partition_size = int(len(trainset) / num_clients)
-            shard_size = int(partition_size / 2)
-            idxs = trainset.targets.argsort()
-            sorted_data = Subset(trainset, idxs)
-            tmp = []
-            for idx in range(num_clients * 2):
-                tmp.append(
-                    Subset(sorted_data, np.arange(shard_size * idx, shard_size * (idx + 1)))
-                )
-            idxs_list = torch.randperm(
-                num_clients * 2, generator=torch.Generator().manual_seed(seed)
-            )
-            datasets = [
-                ConcatDataset((tmp[idxs_list[2 * i]], tmp[idxs_list[2 * i + 1]]))
-                for i in range(num_clients)
-            ]
-
-    return datasets, testset
-
-
-def _balance_classes(
-    trainset: Dataset,
-    seed: Optional[int] = 42,
-) -> Dataset:
-    """Balance the classes of the trainset.
-
-    Trims the dataset so each class contains as many elements as the
-    class that contained the least elements.
-
-    Parameters
-    ----------
-    trainset : Dataset
-        The training dataset that needs to be balanced.
-    seed : int, optional
-        Used to set a fix seed to replicate experiments, by default 42.
-
-    Returns
-    -------
-    Dataset
-        The balanced training dataset.
-    """
-    class_counts = np.bincount(trainset.targets)
-    smallest = np.min(class_counts)
-    idxs = trainset.targets.argsort()
-    tmp = [Subset(trainset, idxs[: int(smallest)])]
-    tmp_targets = [trainset.targets[idxs[: int(smallest)]]]
-    for count in np.cumsum(class_counts):
-        tmp.append(Subset(trainset, idxs[int(count) : int(count + smallest)]))
-        tmp_targets.append(trainset.targets[idxs[int(count) : int(count + smallest)]])
-    unshuffled = ConcatDataset(tmp)
-    unshuffled_targets = torch.cat(tmp_targets)
-    shuffled_idxs = torch.randperm(
-        len(unshuffled), generator=torch.Generator().manual_seed(seed)
-    )
-    shuffled = Subset(unshuffled, shuffled_idxs)
-    shuffled.targets = unshuffled_targets[shuffled_idxs]
-
-    return shuffled
-
-def _sort_by_class(
-    trainset: Dataset,
-) -> Dataset:
-    """Sort dataset by class/label.
-
-    Parameters
-    ----------
-    trainset : Dataset
-        The training dataset that needs to be sorted.
-
-    Returns
-    -------
-    Dataset
-        The sorted training dataset.
-    """
-    class_counts = np.bincount(trainset.targets)
-    idxs = trainset.targets.argsort() # sort targets in ascending order
-
-    tmp = [] # create subset of smallest class
-    tmp_targets = [] # same for targets
-    
-    start = 0
-    for count in np.cumsum(class_counts):
-        tmp.append(Subset(trainset, idxs[start : int(count + start)]))  # add rest of classes
-        tmp_targets.append(trainset.targets[idxs[start : int(count + start)]])
-        start += count
-    sorted_dataset = ConcatDataset(tmp) # concat dataset    
-    sorted_dataset.targets = torch.cat(tmp_targets) # concat targets
-    return sorted_dataset
-
-def _power_law_split(
-    sorted_trainset: Dataset,
-    num_partitions: int,
-    num_labels_per_partition: int = 2,
-    min_data_per_partition: int = 10,
-    mean: float = 0.0,
-    sigma: float = 2.0, 
-) -> Dataset:
-    """Partitions the dataset following a power-law distribution. It
-    follows the implementation of Li et al 2020: https://arxiv.org/abs/1812.06127
-    with default values set accordingtly.
-    
-    Parameters
-    ----------
-    sorted_trainset : Dataset
-        The training dataset sorted by label/class.
-    num_partitions: int
-        Number of partitions to create
-    num_labels_per_partition: int
-        Number of labels to have in each dataset partition. For
-        example if set to two, this means all training examples in
-        a given parition will be long to the same two classes. default 2
-    min_data_per_partition: int
-        Minimum number of datapoints included in each partition, default 10
-    mean: float
-        Mean value for LogNormal distribution to construct power-law, default 0.0
-    sigman: float
-        Sigma value for LogNormal distribution to construct power-law, default 2.0
-
-    Returns
-    -------
-    Dataset
-        The partitioned training dataset.
-    """
-    
-    targets = sorted_trainset.targets
-    full_idx = range(len(targets))
-    
-    class_counts = np.bincount(sorted_trainset.targets)
-    labels_cs = np.cumsum(class_counts)
-    labels_cs = [0] + labels_cs[:-1].tolist()
-    
-    partitions_idx = []
-    num_classes = len(np.bincount(targets))
-    hist = np.zeros(num_classes, dtype=np.int32)
-    
-    # asign min_data_per_partition
-    min_data_per_class = int(min_data_per_partition/num_labels_per_partition)
-    for u_id in range(num_partitions):
-        partitions_idx.append([])
-        for cls_idx in range(num_labels_per_partition):
-            # label for the u_id-th client
-            cls = (u_id + cls_idx) % num_classes
-            # record minimum data
-            indices = list(full_idx[labels_cs[cls] + hist[cls]: labels_cs[cls] + hist[cls] + min_data_per_class])
-            partitions_idx[-1].extend(indices)
-            hist[cls] += min_data_per_class
-
-    # add remaining images following power-law
-    probs = np.random.lognormal(mean, sigma, (num_classes, int(num_partitions/num_classes), num_labels_per_partition))
-    remaining_per_class = class_counts - hist
-    # obtain how many samples each partition should be assigned for each of the labels it contains
-    probs = remaining_per_class.reshape(-1,1,1) * probs/np.sum(probs,(1,2), keepdims=True)
-
-    for u_id in range(num_partitions):
-        for cls_idx in range(num_labels_per_partition):
-            cls = (u_id + cls_idx) % num_classes
-            count = int(probs[cls, u_id//num_classes, cls_idx])
-            
-            # add count of specific class to partition
-            indices = full_idx[labels_cs[cls] + hist[cls]: labels_cs[cls] + hist[cls] + count]
-            partitions_idx[u_id].extend(indices)
-            hist[cls] += count            
-            
-    # construct subsets
-    partitions = [Subset(sorted_trainset, p) for p in partitions_idx]
-    return partitions
